refactor : Group 관련 프론트 파일 및 함수, 변수명 Trip으로 변경
Group 관련 프론트 파일 및 함수, 변수명 Trip으로 변경
