+++ frontend/src/views/trip/TripInvite.vue
