+++ frontend/src/views/trip/TripList.vue
